refactored deck
refactored deck and hand to work with a queue instead of using and left shifting an array.
have a question about the tiles:
in the Balls.java (will rename to player, since it is the player object), i spawn tiles. i want to keep the tiles up so i do not have to click the object everytime i do tile, so my temporary solutoin is using update(), which works, except that the amount of tiles generated is unlimited and i want to limit it until my next action(where i move)